requirements.txt
flet>=0.28.0
cryptography>=44.0.0
argon2-cffi>=23.1.0
